+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Action properties are now properly assigned to transport
 - Parameter schema changed types for "items", "max" and "min"
+- Response reply payload does not leak the call data into the payload anymore
 
 ## [4.0.0] - 2022-03-01
 ### Added